Move tag to atoms
