+++ assets/javascripts/kitten/components/atoms/tag/index.js
