+++ assets/javascripts/kitten/components/cards/backing-card/components/heading-tag.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
-import { Tag } from '../../../../components/tags/tag'
+import { Tag } from '../../../../components/atoms/tag'
 import { DiamondIcon } from '../../../../components/icons/diamond-icon'
 import { StarIcon } from '../../../../components/icons/star-icon'
